update producr controller so Admin users can do everything regular users do, they can also edit/create/remove a produc

+++ Backend/controllers/Products.js
@@ -1,8 +1,12 @@
 const Product = require('../models/Products');
-
+const User = require('../models/Users');
 
 exports.createProduct = async (req, res) => {
   try {
+    const user = await User.findOne({ email: req.user.email });
+    if(user.role !== 'admin') {
+      return res.status(403).json({ message: 'Permission denied' });
+    }
 
     if (!req.body.name || !req.body.price || !req.body.category) {
       return res.status(400).json({
@@ -38,6 +42,11 @@ exports.createProduct = async (req, res) => {
 
 exports.updateProduct = async (req, res) => {
   try {
+    const user = await User.findOne({ email: req.user.email });
+    if(user.role !== 'admin') {
+      return res.status(403).json({ message: 'Permission denied' });
+    }
+
     const { id } = req.params;
     const existingProduct = await Product.findById(id);
     if (!existingProduct) {
@@ -79,6 +88,10 @@ exports.updateProduct = async (req, res) => {
 
 exports.deleteProduct = async (req, res) => {
   try {
+    const user = await User.findOne({ email: req.user.email });
+    if(user.role !== 'admin') {
+      return res.status(403).json({ message: 'Permission denied' });
+    }
     const { id } = req.params;
 
     const deletedProduct = await Product.findByIdAndDelete(id);
